+++ app/controllers/HelloWorldController.scala
@@ -1,0 +1,38 @@
+/*
+ * Copyright 2019 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package controllers
+
+import javax.inject.{Inject, Singleton}
+
+import play.api.mvc._
+import uk.gov.hmrc.play.bootstrap.controller.FrontendController
+import config.AppConfig
+import views.html.hello_world
+
+import scala.concurrent.Future
+
+@Singleton
+class HelloWorldController @Inject()(appConfig: AppConfig, mcc: MessagesControllerComponents)
+    extends FrontendController(mcc) {
+
+  implicit val config: AppConfig = appConfig
+
+  val helloWorld: Action[AnyContent] = Action.async { implicit request =>
+    Future.successful(Ok(hello_world()))
+  }
+
+}
