--- app/uk/gov/hmrc/viewexternalguidancefrontend/config/ErrorHandler.scala
@@ -1,15 +1,0 @@
-package uk.gov.hmrc.viewexternalguidancefrontend.config
-
-import javax.inject.{Inject, Singleton}
-
-import play.api.i18n.MessagesApi
-import play.api.mvc.Request
-import play.twirl.api.Html
-import uk.gov.hmrc.play.bootstrap.http.FrontendErrorHandler
-import uk.gov.hmrc.viewexternalguidancefrontend.views.html.error_template
-
-@Singleton
-class ErrorHandler @Inject()(val messagesApi: MessagesApi, implicit val appConfig: AppConfig) extends FrontendErrorHandler {
-  override def standardErrorTemplate(pageTitle: String, heading: String, message: String)(implicit request: Request[_]): Html =
-    error_template(pageTitle, heading, message)
-}
