--- app/uk/gov/hmrc/viewexternalguidancefrontend/controllers/HelloWorldController.scala
@@ -1,22 +1,0 @@
-package uk.gov.hmrc.viewexternalguidancefrontend.controllers
-
-import javax.inject.{Inject, Singleton}
-
-import play.api.mvc._
-import uk.gov.hmrc.play.bootstrap.controller.FrontendController
-import uk.gov.hmrc.viewexternalguidancefrontend.config.AppConfig
-import uk.gov.hmrc.viewexternalguidancefrontend.views.html.hello_world
-
-import scala.concurrent.Future
-
-@Singleton
-class HelloWorldController @Inject()(appConfig: AppConfig, mcc: MessagesControllerComponents)
-    extends FrontendController(mcc) {
-
-  implicit val config: AppConfig = appConfig
-
-  val helloWorld: Action[AnyContent] = Action.async { implicit request =>
-    Future.successful(Ok(hello_world()))
-  }
-
-}
