--- app/uk/gov/hmrc/viewexternalguidancefrontend/views/error_template.scala.html
@@ -1,12 +1,0 @@
-@import uk.gov.hmrc.viewexternalguidancefrontend.config.AppConfig
-@(pageTitle: String, heading: String, message: String)(implicit request: Request[_], messages: Messages, appConfig: AppConfig)
-
-@contentHeader = {
-  <h1>@heading</h1>
-}
-
-@mainContent = {
-  <p>@message</p>
-}
-
-@govuk_wrapper(appConfig = appConfig, title = pageTitle, contentHeader = Some(contentHeader), mainContent = mainContent)
