--- app/uk/gov/hmrc/viewexternalguidancefrontend/views/hello_world.scala.html
@@ -1,6 +1,0 @@
-@import uk.gov.hmrc.viewexternalguidancefrontend.config.AppConfig
-@()(implicit request: Request[_], messages: Messages, appConfig: AppConfig)
-
-@main_template(title = "Hello from view-external-guidance-frontend", bodyClasses = None) {
-    <h1>Hello from view-external-guidance-frontend !</h1>
-}
