+++ conf/app.routes
@@ -1,3 +1,3 @@
 # microservice specific routes
 
-GET        /hello-world             uk.gov.hmrc.viewexternalguidancefrontend.controllers.HelloWorldController.helloWorld
+GET        /hello-world             controllers.HelloWorldController.helloWorld
