+++ conf/application.conf
@@ -1,3 +1,17 @@
+# Copyright 2019 HM Revenue & Customs
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 include "frontend.conf"
 
 appName = "view-external-guidance-frontend"
@@ -25,7 +39,7 @@ play.http.filters = "uk.gov.hmrc.play.bootstrap.filters.FrontendFilters"
 play.modules.enabled += "uk.gov.hmrc.play.bootstrap.HttpClientModule"
 
 # Custom error handler
-play.http.errorHandler = "uk.gov.hmrc.viewexternalguidancefrontend.config.ErrorHandler"
+play.http.errorHandler = "config.ErrorHandler"
 
 play.filters.headers.contentSecurityPolicy = "default-src 'self' 'unsafe-inline' localhost:9000 localhost:9032 www.google-analytics.com data:"
