+++ conf/prod.routes
@@ -1,5 +1,5 @@
 # Add all the application routes to the app.routes file
-->         /view-external-guidance-frontend              app.Routes
+->         /guidance                  app.Routes
 ->         /                          health.Routes
 ->         /template                  template.Routes
